third_party/blink/web_tests/fast/dom/constructors-cached-navigate-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS testFrame.contentWindow.Image is testFrame.contentWindow.Image
PASS testFrame.contentWindow.Image !== window.Image is true
PASS testFrame.contentWindow.Image.testProperty is "property set successfully"
PASS testFrame.contentWindow.MessageChannel is testFrame.contentWindow.MessageChannel
PASS testFrame.contentWindow.MessageChannel !== window.MessageChannel is true
PASS testFrame.contentWindow.MessageChannel.testProperty is "property set successfully"
PASS testFrame.contentWindow.Option is testFrame.contentWindow.Option
PASS testFrame.contentWindow.Option !== window.Option is true
PASS testFrame.contentWindow.Option.testProperty is "property set successfully"
PASS testFrame.contentWindow.XMLHttpRequest is testFrame.contentWindow.XMLHttpRequest
PASS testFrame.contentWindow.XMLHttpRequest !== window.XMLHttpRequest is true
PASS testFrame.contentWindow.XMLHttpRequest.testProperty is "property set successfully"
PASS testFrame.contentWindow.Audio is testFrame.contentWindow.Audio
PASS testFrame.contentWindow.Audio !== window.Audio is true
PASS testFrame.contentWindow.Audio.testProperty is "property set successfully"
PASS testFrame.contentWindow.Image is testFrame.contentWindow.Image
PASS testFrame.contentWindow.Image !== window.Image is true
PASS testFrame.contentWindow.Image !== storedConstructors.Image is true
PASS storedConstructors.Image.testProperty is "property set successfully"
PASS testFrame.contentWindow.Image.testProperty is undefined.
PASS testFrame.contentWindow.Image.cachedOnOwnerDocument is true
PASS testFrame.contentWindow.MessageChannel is testFrame.contentWindow.MessageChannel
PASS testFrame.contentWindow.MessageChannel !== window.MessageChannel is true
PASS testFrame.contentWindow.MessageChannel !== storedConstructors.MessageChannel is true
PASS storedConstructors.MessageChannel.testProperty is "property set successfully"
PASS testFrame.contentWindow.MessageChannel.testProperty is undefined.
PASS testFrame.contentWindow.MessageChannel.cachedOnOwnerDocument is true
PASS testFrame.contentWindow.Option is testFrame.contentWindow.Option
PASS testFrame.contentWindow.Option !== window.Option is true
PASS testFrame.contentWindow.Option !== storedConstructors.Option is true
PASS storedConstructors.Option.testProperty is "property set successfully"
PASS testFrame.contentWindow.Option.testProperty is undefined.
PASS testFrame.contentWindow.Option.cachedOnOwnerDocument is true
PASS testFrame.contentWindow.XMLHttpRequest is testFrame.contentWindow.XMLHttpRequest
PASS testFrame.contentWindow.XMLHttpRequest !== window.XMLHttpRequest is true
PASS testFrame.contentWindow.XMLHttpRequest !== storedConstructors.XMLHttpRequest is true
PASS storedConstructors.XMLHttpRequest.testProperty is "property set successfully"
PASS testFrame.contentWindow.XMLHttpRequest.testProperty is undefined.
PASS testFrame.contentWindow.XMLHttpRequest.cachedOnOwnerDocument is true
PASS testFrame.contentWindow.Audio is testFrame.contentWindow.Audio
PASS testFrame.contentWindow.Audio !== window.Audio is true
PASS testFrame.contentWindow.Audio !== storedConstructors.Audio is true
PASS storedConstructors.Audio.testProperty is "property set successfully"
PASS testFrame.contentWindow.Audio.testProperty is undefined.
PASS testFrame.contentWindow.Audio.cachedOnOwnerDocument is true
